fix: massive refactor sd extension front end code and logging fix
