+++ aws_extension/cloud_infer_service/utils.py
@@ -1,0 +1,12 @@
+
+class InferManager:
+
+    def run(self, device_id, models_in_cloud, sd_param, is_txt2img) -> str:
+        """
+        Running the actual inference logic to the target
+        :param device_id: the target id
+        :param models_in_cloud: the models used in cloud
+        :param sd_param: the parameter for inference
+        :return: the task id
+        """
+        pass
