+++ aws_extension/sagemaker_ui.py
@@ -14,6 +14,7 @@ from urllib.parse import urljoin
 
 import gradio as gr
 
+import utils
 from modules import shared, scripts
 from modules.ui import create_refresh_button
 from modules.ui_components import FormRow, FormColumn, FormGroup, ToolButton, FormHTML
@@ -29,17 +30,23 @@ import nest_asyncio
 
 from utils import cp, tar, rm
 
+logger = logging.getLogger(__name__)
+logger.setLevel(utils.LOGGING_LEVEL)
+
 inference_job_dropdown = None
 textual_inversion_dropdown = None
 hyperNetwork_dropdown = None
 lora_dropdown = None
 sagemaker_endpoint = None
+modelmerger_merge_on_cloud = None
+interrogate_clip_on_cloud_button = None
+interrogate_deep_booru_on_cloud_button = None
 
 primary_model_name = None
 secondary_model_name = None
 tertiary_model_name = None
 
-#TODO: convert to dynamically init the following variables
+# TODO: convert to dynamically init the following variables
 sagemaker_endpoints = []
 txt2img_inference_job_ids = []
 
@@ -55,7 +62,7 @@ checkpoint_type = ["Stable-diffusion", "embeddings", "Lora", "hypernetworks", "C
 checkpoint_name = ["stable_diffusion", "embeddings", "lora", "hypernetworks", "controlnet", "VAE"]
 stable_diffusion_list = []
 embeddings_list = []
-lora_list = []
+
 hypernetworks_list = []
 controlnet_list = []
 for ckpt_type, ckpt_name in zip(checkpoint_type, checkpoint_name):
@@ -110,10 +117,12 @@ def get_s3_file_names(bucket, folder):
     names = [obj.key for obj in objects]
     return names
 
+
 def get_current_date():
     today = datetime.today()
     formatted_date = today.strftime('%Y-%m-%d')
     return formatted_date
+
 
 def server_request(path):
     api_gateway_url = get_variable_from_json('api_gateway_url')
@@ -129,10 +138,12 @@ def server_request(path):
     response = requests.get(list_endpoint_url, headers=headers)
     return response
 
+
 def datetime_to_short_form(datetime_str):
     dt = datetime.strptime(datetime_str, "%Y-%m-%d %H:%M:%S.%f")
     short_form = dt.strftime("%Y-%m-%d-%H-%M-%S")
     return short_form
+
 
 def update_sagemaker_endpoints():
     global sagemaker_endpoints
@@ -145,7 +156,7 @@ def update_sagemaker_endpoints():
             sagemaker_endpoints.clear()  # Clear the existing list before appending new values
             sagemaker_raw_endpoints = []
             for obj in r:
-                if "EndpointDeploymentJobId" in obj :
+                if "EndpointDeploymentJobId" in obj:
                     if "endpoint_name" in obj:
                         endpoint_name = obj["endpoint_name"]
                         endpoint_status = obj["endpoint_status"]
@@ -166,24 +177,27 @@ def update_sagemaker_endpoints():
                     sagemaker_raw_endpoints.append(endpoint_info)
 
             # Sort the list based on completeTime in descending order
-            sagemaker_endpoints= sorted(sagemaker_raw_endpoints, key=lambda x: x.split('+')[-1], reverse=True)
+            sagemaker_endpoints = sorted(sagemaker_raw_endpoints, key=lambda x: x.split('+')[-1], reverse=True)
 
         else:
-            print("The API response is empty for update_sagemaker_endpoints().")
+            logger.info("The API response is empty for update_sagemaker_endpoints().")
     except Exception as e:
-        print(f"An error occurred while updating SageMaker endpoints: {e}")
+        logger.error(f"An error occurred while updating SageMaker endpoints: {e}")
 
 
-def update_txt2img_inference_job_ids(inference_job_dropdown, txt2img_type_checkbox=True, img2img_type_checkbox=True, interrogate_type_checkbox=True):
+def update_txt2img_inference_job_ids(inference_job_dropdown, txt2img_type_checkbox=True, img2img_type_checkbox=True,
+                                     interrogate_type_checkbox=True):
     # global txt2img_inference_job_ids
     return get_inference_job_list(txt2img_type_checkbox, img2img_type_checkbox, interrogate_type_checkbox)
+
 
 def origin_update_txt2img_inference_job_ids():
     # global origin_txt2img_inference_job_ids
     return query_inference_job_list('', '', '', '', '')
 
 
-def query_page_inference_job_list(task_type: str, status: str, endpoint: str, checkpoint: str, is_img2img: bool, show_all: bool):
+def query_page_inference_job_list(task_type: str, status: str, endpoint: str, checkpoint: str, is_img2img: bool,
+                                  show_all: bool):
     global show_all_inference_job
     show_all_inference_job = show_all
     if is_img2img:
@@ -196,7 +210,7 @@ def query_page_inference_job_list(task_type: str, status: str, endpoint: str, ch
         global img_checkpoint
         img_checkpoint = checkpoint
         opt_type = 'img2img'
-        print(f"{opt_type} {img_task_type} {img_status} {img_endpoint} {img_checkpoint} {show_all}")
+        logger.debug(f"{opt_type} {img_task_type} {img_status} {img_endpoint} {img_checkpoint} {show_all}")
         return query_inference_job_list(task_type, status, endpoint, checkpoint, opt_type)
     else:
         global txt_task_type
@@ -208,7 +222,7 @@ def query_page_inference_job_list(task_type: str, status: str, endpoint: str, ch
         global txt_checkpoint
         txt_checkpoint = checkpoint
         opt_type = 'txt2img'
-        print(f"{opt_type} {txt_task_type} {txt_status} {txt_endpoint} {txt_checkpoint} {show_all}")
+        logger.debug(f"{opt_type} {txt_task_type} {txt_status} {txt_endpoint} {txt_checkpoint} {show_all}")
         return query_inference_job_list(task_type, status, endpoint, checkpoint, opt_type)
 
 
@@ -240,7 +254,7 @@ def query_txt_inference_job_list(task_type: str, status: str, endpoint: str, che
 
 
 def query_inference_job_list(task_type: str, status: str, endpoint: str, checkpoint: str, opt_type: str):
-    print(
+    logger.debug(
         f"query_inference_job_list start！！{status},{task_type},{endpoint},{checkpoint},{start_time_picker_txt_value},{end_time_picker_txt_value} {show_all_inference_job}")
     try:
         body_params = {}
@@ -267,7 +281,7 @@ def query_inference_job_list(task_type: str, status: str, endpoint: str, checkpo
         body_params['limit'] = -1 if show_all_inference_job else 10
         response = server_request_post(f'inference/query-inference-jobs', body_params)
         r = response.json()
-        print(r)
+        logger.debug(r)
         if r:
             txt2img_inference_job_ids.clear()  # Clear the existing list before appending new values
             temp_list = []
@@ -289,11 +303,11 @@ def query_inference_job_list(task_type: str, status: str, endpoint: str, checkpo
             # inference_job_dropdown.update(choices=txt2img_inference_job_ids)
             return gr.Dropdown.update(choices=txt2img_inference_job_ids)
         else:
-            print("The API response is empty.")
+            logger.info("The API response is empty.")
             return gr.Dropdown.update(choices=[])
     except Exception as e:
-        print("Exception occurred when fetching inference_job_ids")
-        print(e)
+        logger.error("Exception occurred when fetching inference_job_ids")
+        logger.error(e)
         return gr.Dropdown.update(choices=[])
 
 
@@ -303,7 +317,7 @@ def get_inference_job_list(txt2img_type_checkbox=True, img2img_type_checkbox=Tru
         txt2img_inference_job_ids.clear()  # Clear the existing list before appending new values
         response = server_request('inference/list-inference-jobs')
         r = response.json()
-        # print(f"response: {response.json()}")
+        logger.debug(f"response: {response.json()}")
         filter_checkbox = False
         selected_types = []
         if txt2img_type_checkbox:
@@ -315,7 +329,7 @@ def get_inference_job_list(txt2img_type_checkbox=True, img2img_type_checkbox=Tru
         if interrogate_type_checkbox:
             selected_types.append('interrogate_deepbooru')
             filter_checkbox = True
-        print(f"selected_types: {selected_types}")
+        logging.debug(f"selected_types: {selected_types}")
         if r:
             txt2img_inference_job_ids.clear()  # Clear the existing list before appending new values
             temp_list = []
@@ -341,19 +355,18 @@ def get_inference_job_list(txt2img_type_checkbox=True, img2img_type_checkbox=Tru
             # inference_job_dropdown.update(choices=txt2img_inference_job_ids)
             return gr.Dropdown.update(choices=txt2img_inference_job_ids)
         else:
-            print("The API response is empty.")
+            logger.debug("The API response is empty.")
             return gr.Dropdown.update(choices=[])
 
     except Exception as e:
-        print("Exception occurred when fetching inference_job_ids")
+        logger.error("Exception occurred when fetching inference_job_ids")
         return gr.Dropdown.update(choices=[])
-
-
 
 
 def get_inference_job(inference_job_id):
     response = server_request(f'inference/get-inference-job?jobID={inference_job_id}')
     return response.json()
+
 
 def get_inference_job_image_output(inference_job_id):
     try:
@@ -365,7 +378,7 @@ def get_inference_job_image_output(inference_job_id):
             txt2img_inference_job_image_list.append(obj_value)
         return txt2img_inference_job_image_list
     except Exception as e:
-        print(f"An error occurred while getting inference job image output: {e}")
+        logger.error(f"An error occurred while getting inference job image output: {e}")
         return []
 
 
@@ -379,7 +392,7 @@ def get_inference_job_param_output(inference_job_id):
             txt2img_inference_job_param_list.append(obj_value)
         return txt2img_inference_job_param_list
     except Exception as e:
-        print(f"An error occurred while getting inference job param output: {e}")
+        logger.error(f"An error occurred while getting inference job param output: {e}")
         return []
 
 
@@ -399,18 +412,17 @@ def download_images(image_urls: list, local_directory: str):
                 f.write(response.content)
             image_list.append(local_path)
         except requests.exceptions.RequestException as e:
-            print(f"Error downloading image {url}: {e}")
+            logger.error(f"Error downloading image {url}: {e}")
     return image_list
 
 
 def get_model_list_by_type(model_type):
-
     api_gateway_url = get_variable_from_json('api_gateway_url')
     api_key = get_variable_from_json('api_token')
 
     # check if api_gateway_url and api_key are set
     if api_gateway_url is None or api_key is None:
-        print("api_gateway_url or api_key is not set")
+        logger.info("api_gateway_url or api_key is not set")
         return []
 
     # Check if api_url ends with '/', if not append it
@@ -446,7 +458,7 @@ def get_model_list_by_type(model_type):
         unique_list = list(set(checkpoint_list))
         return unique_list
     except Exception as e:
-        print(f"Error fetching model list: {e}")
+        logger.error(f"Error fetching model list: {e}")
         return []
 
 
@@ -454,21 +466,26 @@ def update_sd_checkpoints():
     model_type = ["Stable-diffusion"]
     return get_model_list_by_type(model_type)
 
+
 def get_texual_inversion_list():
     model_type = "embeddings"
     return get_model_list_by_type(model_type)
+
 
 def get_lora_list():
     model_type = "Lora"
     return get_model_list_by_type(model_type)
 
+
 def get_hypernetwork_list():
     model_type = "hypernetworks"
     return get_model_list_by_type(model_type)
 
+
 def get_controlnet_model_list():
     model_type = "ControlNet"
     return get_model_list_by_type(model_type)
+
 
 def refresh_all_models():
     api_gateway_url = get_variable_from_json('api_gateway_url')
@@ -478,7 +495,7 @@ def refresh_all_models():
             url = api_gateway_url + f"checkpoints?status=Active&types={rp}"
             response = requests.get(url=url, headers={'x-api-key': api_key})
             json_response = response.json()
-            # print(f"response url json for model {rp} is {json_response}")
+            logger.debug(f"response url json for model {rp} is {json_response}")
             checkpoint_info[rp] = {}
             if "checkpoints" not in json_response.keys():
                 continue
@@ -493,35 +510,37 @@ def refresh_all_models():
                     ckpt_s3_pos = f"{ckpt['s3Location']}/{ckpt_name.split(os.sep)[-1]}"
                     checkpoint_info[ckpt_type][ckpt_name] = ckpt_s3_pos
     except Exception as e:
-        print(f"Error refresh all models: {e}")
+        logger.error(f"Error refresh all models: {e}")
 
 
-def sagemaker_upload_model_s3(sd_checkpoints_path, textual_inversion_path, lora_path, hypernetwork_path, controlnet_model_path, vae_path):
+def sagemaker_upload_model_s3(sd_checkpoints_path, textual_inversion_path, lora_path, hypernetwork_path,
+                              controlnet_model_path, vae_path):
     log = "start upload model to s3:"
 
-    local_paths = [sd_checkpoints_path, textual_inversion_path, lora_path, hypernetwork_path, controlnet_model_path, vae_path]
+    local_paths = [sd_checkpoints_path, textual_inversion_path, lora_path, hypernetwork_path, controlnet_model_path,
+                   vae_path]
 
-    print(f"Refresh checkpoints before upload to get rid of duplicate uploads...")
+    logger.info(f"Refresh checkpoints before upload to get rid of duplicate uploads...")
     refresh_all_models()
 
     for lp, rp in zip(local_paths, checkpoint_type):
         if lp == "" or not lp:
             continue
-        logging.info(f"lp is {lp}")
+        logger.debug(f"lp is {lp}")
         model_name = lp.split(os.sep)[-1]
 
         exist_model_list = list(checkpoint_info[rp].keys())
 
         if model_name in exist_model_list:
-            logging.info(f"!!!skip to upload duplicate model {model_name}")
+            logger.info(f"!!!skip to upload duplicate model {model_name}")
             continue
 
         part_size = 1000 * 1024 * 1024
         file_size = os.stat(lp)
-        parts_number = math.ceil(file_size.st_size/part_size)
-        logging.info('!!!!!!!!!!', file_size, parts_number)
+        parts_number = math.ceil(file_size.st_size / part_size)
+        logger.info('!!!!!!!!!!', file_size, parts_number)
 
-        #local_tar_path = f'{model_name}.tar'
+        # local_tar_path = f'{model_name}.tar'
         local_tar_path = model_name
         payload = {
             "checkpoint_type": rp,
@@ -536,29 +555,29 @@ def sagemaker_upload_model_s3(sd_checkpoints_path, textual_inversion_path, lora_
         if not api_gateway_url.endswith('/'):
             api_gateway_url += '/'
         api_key = get_variable_from_json('api_token')
-        logging.info('!!!!!!api_gateway_url', api_gateway_url)
+        logger.info('!!!!!!api_gateway_url', api_gateway_url)
 
         url = str(api_gateway_url) + "checkpoint"
 
-        logging.info(f"Post request for upload s3 presign url: {url}")
+        logger.debug(f"Post request for upload s3 presign url: {url}")
 
         response = requests.post(url=url, json=payload, headers={'x-api-key': api_key})
 
         try:
             json_response = response.json()
-            # print(f"Response json {json_response}")
+            logger.debug(f"Response json {json_response}")
             s3_base = json_response["checkpoint"]["s3_location"]
             checkpoint_id = json_response["checkpoint"]["id"]
-            logging.info(f"Upload to S3 {s3_base}")
-            logging.info(f"Checkpoint ID: {checkpoint_id}")
+            logger.debug(f"Upload to S3 {s3_base}")
+            logger.debug(f"Checkpoint ID: {checkpoint_id}")
 
             s3_signed_urls_resp = json_response["s3PresignUrl"][local_tar_path]
             # Upload src model to S3.
-            if rp != "embeddings" :
+            if rp != "embeddings":
                 local_model_path_in_repo = os.sep.join(['models', rp, model_name])
             else:
                 local_model_path_in_repo = os.sep.join([rp, model_name])
-            logging.info("Pack the model file.")
+            logger.debug("Pack the model file.")
             cp(lp, local_model_path_in_repo, recursive=True)
             if rp == "Stable-diffusion":
                 model_yaml_name = model_name.split('.')[0] + ".yaml"
@@ -566,7 +585,8 @@ def sagemaker_upload_model_s3(sd_checkpoints_path, textual_inversion_path, lora_
                 local_model_yaml_path_in_repo = os.sep.join(["models", rp, model_yaml_name])
                 if os.path.isfile(local_model_yaml_path):
                     cp(local_model_yaml_path, local_model_yaml_path_in_repo, recursive=True)
-                    tar(mode='c', archive=local_tar_path, sfiles=[local_model_path_in_repo, local_model_yaml_path_in_repo], verbose=True)
+                    tar(mode='c', archive=local_tar_path,
+                        sfiles=[local_model_path_in_repo, local_model_yaml_path_in_repo], verbose=True)
                 else:
                     tar(mode='c', archive=local_tar_path, sfiles=[local_model_path_in_repo], verbose=True)
             else:
@@ -586,16 +606,16 @@ def sagemaker_upload_model_s3(sd_checkpoints_path, textual_inversion_path, lora_
             # Start creating model on cloud.
             response = requests.put(url=url, json=payload, headers={'x-api-key': api_key})
             s3_input_path = s3_base
-            logging.info(response)
+            logger.debug(response)
 
             log = f"\n finish upload {local_tar_path} to {s3_base}"
 
             # os.system(f"rm {local_tar_path}")
             rm(local_tar_path, recursive=True)
         except Exception as e:
-            logging.info(f"fail to upload model {lp}, error: {e}")
+            logger.error(f"fail to upload model {lp}, error: {e}")
 
-    logging.info(f"Refresh checkpoints after upload...")
+    logger.debug(f"Refresh checkpoints after upload...")
     refresh_all_models()
     return log, None, None, None, None, None, None
 
@@ -606,10 +626,11 @@ def sagemaker_upload_model_s3_local():
 
 
 def generate_on_cloud(sagemaker_endpoint):
-    logging.info(f"checkpiont_info {checkpoint_info}")
-    logging.info(f"sagemaker endpoint {sagemaker_endpoint}")
+    logger.info(f"checkpiont_info {checkpoint_info}")
+    logger.info(f"sagemaker endpoint {sagemaker_endpoint}")
     text = "failed to check endpoint"
     return plaintext_to_html(text)
+
 
 # create a global event loop and apply the patch to allow nested event loops in single thread
 loop = asyncio.get_event_loop()
@@ -617,13 +638,14 @@ nest_asyncio.apply()
 
 MAX_RUNNING_LIMIT = 10
 
+
 def async_loop_wrapper(f):
     global loop
     # check if there are any running or queued tasks inside the event loop
     if loop.is_running():
         # Calculate the number of running tasks
         while len([task for task in asyncio.all_tasks(loop) if not task.done()]) > MAX_RUNNING_LIMIT:
-            logging.info(f'Waiting for {MAX_RUNNING_LIMIT} running tasks to complete')
+            logger.debug(f'Waiting for {MAX_RUNNING_LIMIT} running tasks to complete')
             time.sleep(1)
     else:
         # check if loop is closed and create a new one
@@ -631,11 +653,12 @@ def async_loop_wrapper(f):
             loop = asyncio.new_event_loop()
             asyncio.set_event_loop(loop)
             # log this event since it should never happen
-            logging.info('Event loop was closed, created a new one')
+            logger.debug('Event loop was closed, created a new one')
 
     # Add new task to the event loop
     result = loop.run_until_complete(f())
     return result
+
 
 def async_loop_wrapper_with_input(sagemaker_endpoint, type):
     global loop
@@ -643,7 +666,7 @@ def async_loop_wrapper_with_input(sagemaker_endpoint, type):
     if loop.is_running():
         # Calculate the number of running tasks
         while len([task for task in asyncio.all_tasks(loop) if not task.done()]) > MAX_RUNNING_LIMIT:
-            logging.info(f'Waiting for {MAX_RUNNING_LIMIT} running tasks to complete')
+            logger.debug(f'Waiting for {MAX_RUNNING_LIMIT} running tasks to complete')
             time.sleep(1)
     else:
         # check if loop is closed and create a new one
@@ -651,27 +674,35 @@ def async_loop_wrapper_with_input(sagemaker_endpoint, type):
             loop = asyncio.new_event_loop()
             asyncio.set_event_loop(loop)
             # log this event since it should never happen
-            logging.info('Event loop was closed, created a new one')
+            logger.debug('Event loop was closed, created a new one')
 
     # Add new task to the event loop
     result = loop.run_until_complete(call_remote_inference(sagemaker_endpoint, type))
     return result
 
+
 def call_txt2img_inference(sagemaker_endpoint):
     return async_loop_wrapper_with_input(sagemaker_endpoint, 'txt2img')
 
-def call_img2img_inference(sagemaker_endpoint, init_img, sketch, init_img_with_mask, inpaint_color_sketch, init_img_inpaint, init_mask_inpaint):
+
+def call_img2img_inference(sagemaker_endpoint, init_img, sketch, init_img_with_mask, inpaint_color_sketch,
+                           init_img_inpaint, init_mask_inpaint):
     return async_loop_wrapper_with_input(sagemaker_endpoint, 'img2img')
 
-def call_interrogate_clip(sagemaker_endpoint, init_img, sketch, init_img_with_mask, inpaint_color_sketch, init_img_inpaint, init_mask_inpaint):
+
+def call_interrogate_clip(sagemaker_endpoint, init_img, sketch, init_img_with_mask, inpaint_color_sketch,
+                          init_img_inpaint, init_mask_inpaint):
     return async_loop_wrapper_with_input(sagemaker_endpoint, 'interrogate_clip')
 
-def call_interrogate_deepbooru(sagemaker_endpoint, init_img, sketch, init_img_with_mask, inpaint_color_sketch, init_img_inpaint, init_mask_inpaint):
+
+def call_interrogate_deepbooru(sagemaker_endpoint, init_img, sketch, init_img_with_mask, inpaint_color_sketch,
+                               init_img_inpaint, init_mask_inpaint):
     return async_loop_wrapper_with_input(sagemaker_endpoint, 'interrogate_deepbooru')
 
+
 async def call_remote_inference(sagemaker_endpoint, type):
-    logging.info(f"chosen ep {sagemaker_endpoint}")
-    logging.info(f"inference type is {type}")
+    logger.debug(f"chosen ep {sagemaker_endpoint}")
+    logger.debug(f"inference type is {type}")
 
     if sagemaker_endpoint == '':
         image_list = []  # Return an empty list if selected_value is None
@@ -706,22 +737,22 @@ async def call_remote_inference(sagemaker_endpoint, type):
     checkpoint_info['sagemaker_endpoint'] = sagemaker_endpoint.split("+")[0]
     payload = checkpoint_info
     payload['task_type'] = type
-    logging.info(f"checkpointinfo is {payload}")
+    logger.info(f"checkpointinfo is {payload}")
 
     inference_url = f"{api_gateway_url}inference/run-sagemaker-inference"
     response = requests.post(inference_url, json=payload, headers=headers)
-    logging.info(f"Raw server response: {response.text}")
+    logger.info(f"Raw server response: {response.text}")
     try:
         r = response.json()
     except JSONDecodeError as e:
-        logging.info(f"Failed to decode JSON response: {e}")
-        logging.info(f"Raw server response: {response.text}")
+        logger.error(f"Failed to decode JSON response: {e}")
+        logger.error(f"Raw server response: {response.text}")
     else:
         inference_id = r.get('inference_id')  # Assuming the response contains 'inference_id' field
         try:
             return process_result_by_inference_id(inference_id)
         except Exception as e:
-            logging.info(f"Failed to get inference job {inference_id}, error: {e}")
+            logger.error(f"Failed to get inference job {inference_id}, error: {e}")
 
 
 def process_result_by_inference_id(inference_id):
@@ -730,7 +761,6 @@ def process_result_by_inference_id(inference_id):
     infotexts = f"Inference id is {inference_id}, please check all historical inference result in 'Inference Job' dropdown list"
     json_list = []
     prompt_txt = ''
-
 
     resp = get_inference_job(inference_id)
     if resp is None:
@@ -752,12 +782,14 @@ def process_result_by_inference_id(inference_id):
                 inference_param_json_list = get_inference_job_param_output(inference_id)
                 # todo: these not need anymore
                 if resp['taskType'] == "txt2img":
-                    image_list = download_images(images,f"outputs/txt2img-images/{get_current_date()}/{inference_id}/")
-                    json_list = download_images(inference_param_json_list, f"outputs/txt2img-images/{get_current_date()}/{inference_id}/")
+                    image_list = download_images(images, f"outputs/txt2img-images/{get_current_date()}/{inference_id}/")
+                    json_list = download_images(inference_param_json_list,
+                                                f"outputs/txt2img-images/{get_current_date()}/{inference_id}/")
                     json_file = f"outputs/txt2img-images/{get_current_date()}/{inference_id}/{inference_id}_param.json"
                 elif resp['taskType'] == "img2img":
-                    image_list = download_images(images,f"outputs/img2img-images/{get_current_date()}/{inference_id}/")
-                    json_list = download_images(inference_param_json_list, f"outputs/img2img-images/{get_current_date()}/{inference_id}/")
+                    image_list = download_images(images, f"outputs/img2img-images/{get_current_date()}/{inference_id}/")
+                    json_list = download_images(inference_param_json_list,
+                                                f"outputs/img2img-images/{get_current_date()}/{inference_id}/")
                     json_file = f"outputs/img2img-images/{get_current_date()}/{inference_id}/{inference_id}_param.json"
                 if os.path.isfile(json_file):
                     with open(json_file) as f:
@@ -765,25 +797,26 @@ def process_result_by_inference_id(inference_id):
                         info_text = log_file["info"]
                         infotexts = f"Inference id is {inference_id}\n" + json.loads(info_text)["infotexts"][0]
                 else:
-                    logging.info(f"File {json_file} does not exist.")
+                    logger.debug(f"File {json_file} does not exist.")
                     info_text = 'something wrong when trying to download the inference parameters'
                     infotexts = info_text
                 return image_list, info_text, plaintext_to_html(infotexts)
         else:
             return image_list, info_text, plaintext_to_html(infotexts)
 
+
 def sagemaker_endpoint_delete(delete_endpoint_list):
-    logging.info(f"start delete sagemaker endpoint delete function")
-    logging.info(f"delete endpoint list: {delete_endpoint_list}")
+    logger.debug(f"start delete sagemaker endpoint delete function")
+    logger.debug(f"delete endpoint list: {delete_endpoint_list}")
     api_gateway_url = get_variable_from_json('api_gateway_url')
     api_key = get_variable_from_json('api_token')
 
     delete_endpoint_list = [item.split('+')[0] for item in delete_endpoint_list]
-    logging.info(f"delete endpoint list: {delete_endpoint_list}")
+    logger.debug(f"delete endpoint list: {delete_endpoint_list}")
 
     # check if api_gateway_url and api_key are set
     if api_gateway_url is None or api_key is None:
-        logging.info("api_gateway_url and api_key are not set")
+        logger.debug("api_gateway_url and api_key are not set")
         return
 
     # Check if api_url ends with '/', if not append it
@@ -804,9 +837,10 @@ def sagemaker_endpoint_delete(delete_endpoint_list):
     try:
         response = requests.post(deployment_url, json=payload, headers=headers)
         r = response.json()
-        logging.info(f"response for rest api {r}")
+        logger.debug(f"response for rest api {r}")
         return r
     except Exception as e:
+        logger.error(e)
         return f"Failed to delete sagemaker endpoint with exception: {e}"
 
 
@@ -819,14 +853,15 @@ def sagemaker_deploy(endpoint_name_textbox, instance_type, initial_instance_coun
         (None)
     """
     # function code to call sagemaker deploy api
-    logging.info(f"start deploying instance type: {instance_type} with count {initial_instance_count} with autoscaling {autoscaling_enabled}............")
+    logger.debug(
+        f"start deploying instance type: {instance_type} with count {initial_instance_count} with autoscaling {autoscaling_enabled}............")
 
     api_gateway_url = get_variable_from_json('api_gateway_url')
     api_key = get_variable_from_json('api_token')
 
     # check if api_gateway_url and api_key are set
     if api_gateway_url is None or api_key is None:
-        logging.info("api_gateway_url and api_key are not set")
+        logger.debug("api_gateway_url and api_key are not set")
         return
     # Check if api_url ends with '/', if not append it
     if not api_gateway_url.endswith('/'):
@@ -849,13 +884,15 @@ def sagemaker_deploy(endpoint_name_textbox, instance_type, initial_instance_coun
     try:
         response = requests.post(deployment_url, json=payload, headers=headers)
         r = response.json()
-        logging.info(f"response for rest api {r}")
+        logger.debug(f"response for rest api {r}")
         return "Endpoint deployment started"
     except Exception as e:
+        logger.error(e)
         return f"Failed to start endpoint deployment with exception: {e}"
 
+
 def modelmerger_on_cloud_func(primary_model_name, secondary_model_name, teritary_model_name):
-    logging.info(f"function under development, current checkpoint_info is {checkpoint_info}")
+    logger.debug(f"function under development, current checkpoint_info is {checkpoint_info}")
     api_gateway_url = get_variable_from_json('api_gateway_url')
     # Check if api_url ends with '/', if not append it
     if not api_gateway_url.endswith('/'):
@@ -863,14 +900,14 @@ def modelmerger_on_cloud_func(primary_model_name, secondary_model_name, teritary
     api_key = get_variable_from_json('api_token')
 
     if api_gateway_url is None:
-        logging.info(f"modelmerger: failed to get the api-gateway url, can not fetch remote data")
+        logger.debug(f"modelmerger: failed to get the api-gateway url, can not fetch remote data")
         return []
     modelmerge_url = f"{api_gateway_url}inference/run-model-merge"
 
     payload = {
-        "primary_model_name" : primary_model_name,
-        "secondary_model_name" : secondary_model_name,
-        "tertiary_model_name" : teritary_model_name
+        "primary_model_name": primary_model_name,
+        "secondary_model_name": secondary_model_name,
+        "tertiary_model_name": teritary_model_name
     }
 
     headers = {
@@ -883,17 +920,18 @@ def modelmerger_on_cloud_func(primary_model_name, secondary_model_name, teritary
     try:
         r = response.json()
     except JSONDecodeError as e:
-        logging.info(f"Failed to decode JSON response: {e}")
-        logging.info(f"Raw server response: {response.text}")
+        logger.error(f"Failed to decode JSON response: {e}")
+        logger.error(f"Raw server response: {response.text}")
     else:
-        logging.info(f"response for rest api {r}")
+        logger.debug(f"response for rest api {r}")
+
 
 # def txt2img_config_save():
 #     # placeholder for saving txt2img config
 #     pass
 
 def displayEndpointInfo(input_string: str):
-    logging.info(f"selected value is {input_string}")
+    logger.debug(f"selected value is {input_string}")
     if not input_string:
         return
     parts = input_string.split('+')
@@ -914,20 +952,25 @@ def displayEndpointInfo(input_string: str):
     else:
         return plaintext_to_html("")
 
+
 def update_txt2imgPrompt_from_TextualInversion(selected_items, txt2img_prompt):
     return update_txt2imgPrompt_from_model_select(selected_items, txt2img_prompt, 'embeddings', False)
+
 
 def update_txt2imgPrompt_from_Hypernetworks(selected_items, txt2img_prompt):
     return update_txt2imgPrompt_from_model_select(selected_items, txt2img_prompt, 'hypernetworks', True)
 
+
 def update_txt2imgPrompt_from_Lora(selected_items, txt2img_prompt):
     return update_txt2imgPrompt_from_model_select(selected_items, txt2img_prompt, 'Lora', True)
 
-def update_txt2imgPrompt_from_model_select(selected_items, txt2img_prompt, model_name='embeddings', with_angle_brackets=False):
-    logging.info(selected_items) #example ['FastNegativeV2.pt']
-    logging.info(txt2img_prompt)
-    logging.info(get_model_list_by_type('embeddings'))
-    full_dropdown_items = get_model_list_by_type(model_name) #example ['FastNegativeV2.pt', 'okuryl3nko.pt']
+
+def update_txt2imgPrompt_from_model_select(selected_items, txt2img_prompt, model_name='embeddings',
+                                           with_angle_brackets=False):
+    logger.debug(selected_items)  # example ['FastNegativeV2.pt']
+    logger.debug(txt2img_prompt)
+    logger.debug(get_model_list_by_type('embeddings'))
+    full_dropdown_items = get_model_list_by_type(model_name)  # example ['FastNegativeV2.pt', 'okuryl3nko.pt']
 
     # Remove extensions from selected_items and full_dropdown_items
     selected_items = [item.split('.')[0] for item in selected_items]
@@ -959,8 +1002,8 @@ def update_txt2imgPrompt_from_model_select(selected_items, txt2img_prompt, model
 
 
 def fake_gan(selected_value, original_prompt):
-    logging.info(f"selected value is {selected_value}")
-    logging.info(f"original prompt is {original_prompt}")
+    logger.debug(f"selected value is {selected_value}")
+    logger.debug(f"original prompt is {original_prompt}")
     if selected_value is not None:
         delimiter = "-->"
         parts = selected_value.split(delimiter)
@@ -979,27 +1022,29 @@ def fake_gan(selected_value, original_prompt):
             inference_pram_json_list = get_inference_job_param_output(inference_job_id)
             # output directory mapping to task type
             if inference_job_taskType == "txt2img":
-                image_list = download_images(images,f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
-                json_list = download_images(inference_pram_json_list, f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
+                image_list = download_images(images, f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
+                json_list = download_images(inference_pram_json_list,
+                                            f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
                 json_file = f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/{inference_job_id}_param.json"
             elif inference_job_taskType == "img2img":
-                image_list = download_images(images,f"outputs/img2img-images/{get_current_date()}/{inference_job_id}/")
-                json_list = download_images(inference_pram_json_list, f"outputs/img2img-images/{get_current_date()}/{inference_job_id}/")
+                image_list = download_images(images, f"outputs/img2img-images/{get_current_date()}/{inference_job_id}/")
+                json_list = download_images(inference_pram_json_list,
+                                            f"outputs/img2img-images/{get_current_date()}/{inference_job_id}/")
                 json_file = f"outputs/img2img-images/{get_current_date()}/{inference_job_id}/{inference_job_id}_param.json"
-            logging.info(f"{str(images)}")
-            logging.info(f"{str(inference_pram_json_list)}")
+            logger.debug(f"{str(images)}")
+            logger.debug(f"{str(inference_pram_json_list)}")
             if os.path.isfile(json_file):
                 with open(json_file) as f:
                     log_file = json.load(f)
                     info_text = log_file["info"]
                     infotexts = json.loads(info_text)["infotexts"][0]
             else:
-                logging.info(f"File {json_file} does not exist.")
+                logger.debug(f"File {json_file} does not exist.")
                 info_text = 'something wrong when trying to download the inference parameters'
                 infotexts = 'something wrong when trying to download the inference parameters'
         elif inference_job_taskType in ["interrogate_clip", "interrogate_deepbooru"]:
             job_status = get_inference_job(inference_job_id)
-            logging.info(job_status)
+            logger.debug(job_status)
             caption = job_status['caption']
             prompt_txt = caption
             image_list = []  # Return an empty list if selected_value is None
@@ -1014,21 +1059,23 @@ def fake_gan(selected_value, original_prompt):
         infotexts = ''
     return image_list, info_text, plaintext_to_html(infotexts), prompt_txt
 
-def display_inference_result(inference_id: str ):
-    logging.info(f"selected value is {inference_id}")
+
+def display_inference_result(inference_id: str):
+    logger.debug(f"selected value is {inference_id}")
     if inference_id is not None:
         # Extract the InferenceJobId value
         inference_job_id = inference_id
         images = get_inference_job_image_output(inference_job_id)
         image_list = []
-        image_list = download_images(images,f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
+        image_list = download_images(images, f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
 
         inference_pram_json_list = get_inference_job_param_output(inference_job_id)
         json_list = []
-        json_list = download_images(inference_pram_json_list, f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
+        json_list = download_images(inference_pram_json_list,
+                                    f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/")
 
-        logging.info(f"{str(images)}")
-        logging.info(f"{str(inference_pram_json_list)}")
+        logger.debug(f"{str(images)}")
+        logger.debug(f"{str(inference_pram_json_list)}")
 
         json_file = f"outputs/txt2img-images/{get_current_date()}/{inference_job_id}/{inference_job_id}_param.json"
 
@@ -1046,8 +1093,9 @@ def display_inference_result(inference_id: str ):
 
     return image_list, info_text, plaintext_to_html(infotexts)
 
+
 def init_refresh_resource_list_from_cloud():
-    logging.info(f"start refreshing resource list from cloud")
+    logger.debug(f"start refreshing resource list from cloud")
     if get_variable_from_json('api_gateway_url') is not None:
         update_sagemaker_endpoints()
         refresh_all_models()
@@ -1057,11 +1105,11 @@ def init_refresh_resource_list_from_cloud():
         get_controlnet_model_list()
         get_inference_job_list()
     else:
-        logging.info(f"there is no api-gateway url and token in local file,")
+        logger.debug(f"there is no api-gateway url and token in local file,")
 
 
 def on_txt_time_change(start_time, end_time):
-    logging.info(f"!!!!!!!!!on_txt_time_change!!!!!!{start_time},{end_time}")
+    logger.debug(f"!!!!!!!!!on_txt_time_change!!!!!!{start_time},{end_time}")
     global start_time_picker_txt_value
     global end_time_picker_txt_value
     start_time_picker_txt_value = start_time
@@ -1070,7 +1118,7 @@ def on_txt_time_change(start_time, end_time):
 
 
 def on_img_time_change(start_time, end_time):
-    logging.info(f"!!!!!!!!!on_img_time_change!!!!!!{start_time},{end_time}")
+    logger.debug(f"!!!!!!!!!on_img_time_change!!!!!!{start_time},{end_time}")
     global start_time_picker_img_value
     global end_time_picker_img_value
     start_time_picker_img_value = start_time
@@ -1095,8 +1143,9 @@ def create_ui(is_img2img):
                                                  elem_id="sagemaker_endpoint_dropdown"
                                                  )
 
-                modules.ui.create_refresh_button(sagemaker_endpoint, update_sagemaker_endpoints, lambda: {"choices": sagemaker_endpoints, "value": None}, "refresh_sagemaker_endpoints")
-
+                modules.ui.create_refresh_button(sagemaker_endpoint, update_sagemaker_endpoints,
+                                                 lambda: {"choices": sagemaker_endpoints, "value": None},
+                                                 "refresh_sagemaker_endpoints")
 
                 # with gr.Row():
                 #     sd_checkpoint = gr.Dropdown(multiselect=True, label="Stable Diffusion Checkpoint", choices=sorted(update_sd_checkpoints()), elem_id="stable_diffusion_checkpoint_dropdown")
@@ -1111,9 +1160,11 @@ def create_ui(is_img2img):
                 if is_img2img:
                     with gr.Row():
                         with gr.Column():
-                            interrogate_clip_on_cloud_button = gr.Button(value="Interrogate CLIP", elem_id="interrogate_clip_on_cloud_button")
+                            interrogate_clip_on_cloud_button = gr.Button(value="Interrogate CLIP",
+                                                                         elem_id="interrogate_clip_on_cloud_button")
                         with gr.Column():
-                            interrogate_deep_booru_on_cloud_button = gr.Button(value="Interrogte DeepBooru", elem_id="interrogate_deep_booru_on_cloud_button")
+                            interrogate_deep_booru_on_cloud_button = gr.Button(value="Interrogte DeepBooru",
+                                                                               elem_id="interrogate_deep_booru_on_cloud_button")
                         # with gr.Column():
                         #     generate_on_cloud_button_with_js_img2img = gr.Button(value="Generate on Cloud", variant='primary', elem_id="generate_on_cloud_with_cloud_config_button_img2img",queue=True, show_progress=True)
             with gr.Row():
@@ -1124,8 +1175,13 @@ def create_ui(is_img2img):
                                                      label="Inference Job: Time-Type-Status-Uuid",
                                                      elem_id="txt2img_inference_job_ids_dropdown"
                                                      )
-                txt2img_inference_job_ids_refresh_button = modules.ui.create_refresh_button(inference_job_dropdown, origin_update_txt2img_inference_job_ids, lambda: {"choices": txt2img_inference_job_ids, "value": None}, "refresh_txt2img_inference_job_ids")
-                # logging.info(f"gr.Row() txt2img_inference_job_ids is {txt2img_inference_job_ids}")
+                txt2img_inference_job_ids_refresh_button = modules.ui.create_refresh_button(inference_job_dropdown,
+                                                                                            origin_update_txt2img_inference_job_ids,
+                                                                                            lambda: {
+                                                                                                "choices": txt2img_inference_job_ids,
+                                                                                                "value": None},
+                                                                                            "refresh_txt2img_inference_job_ids")
+                # logger.debug(f"gr.Row() txt2img_inference_job_ids is {txt2img_inference_job_ids}")
 
             # with gr.Row():
             #     with gr.Column(scale=1):
@@ -1151,15 +1207,23 @@ def create_ui(is_img2img):
                 with gr.Column(scale=2):
                     with gr.Row():
                         task_type_choices = ["txt2img", "img2img", "interrogate_clip", "interrogate_deepbooru"]
-                        task_type_dropdown = gr.Dropdown(label="Task Type", choices=task_type_choices, elem_id="task_type_ids_dropdown")
+                        task_type_dropdown = gr.Dropdown(label="Task Type", choices=task_type_choices,
+                                                         elem_id="task_type_ids_dropdown")
                         status_choices = ["succeed", "inprogress", "failed"]
-                        status_dropdown = gr.Dropdown(label="Status", choices=status_choices, elem_id="task_status_dropdown")
+                        status_dropdown = gr.Dropdown(label="Status", choices=status_choices,
+                                                      elem_id="task_status_dropdown")
                     with gr.Row():
-                        sagemaker_endpoint_filter = gr.Dropdown(sagemaker_endpoints, label="SageMaker Endpoint", elem_id="sagemaker_endpoint_dropdown" )
-                        modules.ui.create_refresh_button(sagemaker_endpoint_filter, update_sagemaker_endpoints, lambda: {"choices": sagemaker_endpoints}, "refresh_sagemaker_endpoints")
+                        sagemaker_endpoint_filter = gr.Dropdown(sagemaker_endpoints, label="SageMaker Endpoint",
+                                                                elem_id="sagemaker_endpoint_dropdown")
+                        modules.ui.create_refresh_button(sagemaker_endpoint_filter, update_sagemaker_endpoints,
+                                                         lambda: {"choices": sagemaker_endpoints},
+                                                         "refresh_sagemaker_endpoints")
                     with gr.Row():
-                        sd_checkpoint_filter = gr.Dropdown(label="Checkpoint", choices=sorted(update_sd_checkpoints()), elem_id="stable_diffusion_checkpoint_dropdown")
-                        modules.ui.create_refresh_button(sd_checkpoint_filter, update_sd_checkpoints, lambda: { "choices": sorted(update_sd_checkpoints())}, "refresh_sd_checkpoints")
+                        sd_checkpoint_filter = gr.Dropdown(label="Checkpoint", choices=sorted(update_sd_checkpoints()),
+                                                           elem_id="stable_diffusion_checkpoint_dropdown")
+                        modules.ui.create_refresh_button(sd_checkpoint_filter, update_sd_checkpoints,
+                                                         lambda: {"choices": sorted(update_sd_checkpoints())},
+                                                         "refresh_sd_checkpoints")
                     if is_img2img:
                         with gr.Row():
                             start_time_picker_img = gr.HTML(elem_id="start_timepicker_img_e",
@@ -1291,19 +1355,32 @@ def create_ui(is_img2img):
             sagemaker_html_log = gr.HTML(elem_id=f'html_log_sagemaker')
             with FormRow(elem_id="modelmerger_models_in_the_cloud"):
                 global primary_model_name
-                primary_model_name = gr.Dropdown(choices=sorted(update_sd_checkpoints()), elem_id="modelmerger_primary_model_name_in_the_cloud", label="Primary model (A) in the cloud")
-                create_refresh_button(primary_model_name, update_sd_checkpoints, lambda: {"choices": sorted(update_sd_checkpoints())}, "refresh_checkpoint_A_in_the_cloud")
+                primary_model_name = gr.Dropdown(choices=sorted(update_sd_checkpoints()),
+                                                 elem_id="modelmerger_primary_model_name_in_the_cloud",
+                                                 label="Primary model (A) in the cloud")
+                create_refresh_button(primary_model_name, update_sd_checkpoints,
+                                      lambda: {"choices": sorted(update_sd_checkpoints())},
+                                      "refresh_checkpoint_A_in_the_cloud")
 
                 global secondary_model_name
-                secondary_model_name = gr.Dropdown(choices=sorted(update_sd_checkpoints()), elem_id="modelmerger_secondary_model_name_in_the_cloud", label="Secondary model (B) in the cloud")
-                create_refresh_button(secondary_model_name, update_sd_checkpoints, lambda: {"choices": sorted(update_sd_checkpoints())}, "refresh_checkpoint_B_in_the_cloud")
+                secondary_model_name = gr.Dropdown(choices=sorted(update_sd_checkpoints()),
+                                                   elem_id="modelmerger_secondary_model_name_in_the_cloud",
+                                                   label="Secondary model (B) in the cloud")
+                create_refresh_button(secondary_model_name, update_sd_checkpoints,
+                                      lambda: {"choices": sorted(update_sd_checkpoints())},
+                                      "refresh_checkpoint_B_in_the_cloud")
 
                 global tertiary_model_name
-                tertiary_model_name = gr.Dropdown(choices=sorted(update_sd_checkpoints()), elem_id="modelmerger_tertiary_model_name_in_the_cloud", label="Tertiary model (C) in the cloud")
-                create_refresh_button(tertiary_model_name, update_sd_checkpoints, lambda: {"choices": sorted(update_sd_checkpoints())}, "refresh_checkpoint_C_in_the_cloud")
+                tertiary_model_name = gr.Dropdown(choices=sorted(update_sd_checkpoints()),
+                                                  elem_id="modelmerger_tertiary_model_name_in_the_cloud",
+                                                  label="Tertiary model (C) in the cloud")
+                create_refresh_button(tertiary_model_name, update_sd_checkpoints,
+                                      lambda: {"choices": sorted(update_sd_checkpoints())},
+                                      "refresh_checkpoint_C_in_the_cloud")
             with gr.Row():
                 global modelmerger_merge_on_cloud
-                modelmerger_merge_on_cloud = gr.Button(elem_id="modelmerger_merge_in_the_cloud", value="Merge on Cloud", variant='primary')
+                modelmerger_merge_on_cloud = gr.Button(elem_id="modelmerger_merge_in_the_cloud", value="Merge on Cloud",
+                                                       variant='primary')
 
     # return sagemaker_endpoint, sd_checkpoint, sd_checkpoint_refresh_button, textual_inversion_dropdown, lora_dropdown, hyperNetwork_dropdown, controlnet_dropdown, inference_job_dropdown, txt2img_inference_job_ids_refresh_button, primary_model_name, secondary_model_name, tertiary_model_name, modelmerger_merge_on_cloud
     return sagemaker_endpoint, inference_job_dropdown, txt2img_inference_job_ids_refresh_button, primary_model_name, secondary_model_name, tertiary_model_name, modelmerger_merge_on_cloud
